+++ src/Wiegand.h
@@ -12,24 +12,24 @@ class WIEGAND {
 public:
 	WIEGAND();
 	void begin();
-	void begin(int pinD0, int pinD1);
-	void begin(int pinD0, int pinIntD0, int pinD1, int pinIntD1);
+	void begin(uint8_t pinD0, uint8_t pinD1);
+	void begin(uint8_t pinD0, uint8_t pinIntD0, uint8_t pinD1, uint8_t pinIntD1);
 	bool available();
-	unsigned long getCode();
-	int getWiegandType();
+	uint32_t  getCode();
+	uint8_t getWiegandType();
 	
 private:
 	static void ReadD0();
 	static void ReadD1();
 	static bool DoWiegandConversion ();
-	static unsigned long GetCardId (volatile unsigned long *codehigh, volatile unsigned long *codelow, char bitlength);
+	static uint32_t  GetCardId (volatile uint32_t  *codehigh, volatile uint32_t  *codelow, char bitlength);
 	
-	static volatile unsigned long 	_cardTempHigh;
-	static volatile unsigned long 	_cardTemp;
-	static volatile unsigned long 	_lastWiegand;
-	static volatile int				_bitCount;	
-	static int				_wiegandType;
-	static unsigned long	_code;
+	static volatile uint32_t  	_cardTempHigh;
+	static volatile uint32_t  	_cardTemp;
+	static volatile uint32_t  	_lastWiegand;
+	static volatile uint8_t		_bitCount;
+	static uint8_t				_wiegandType;
+	static uint32_t 	_code;
 };
 
 #endif
